requirement.txt
pillow==10.4.0
pygame==2.6.0
tk==0.1.0